pyproject.toml
[project]
name = "agent-book"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy<2.0.0",
    "ipykernel>=6.29.5",
    "hydra-core>=1.3.2",
    "openai==1.40.6",
    "langchain-core==0.3.0",
    "langchain-community==0.3.0",
    "langchain-text-splitters==0.3.0",
    "langchain-anthropic==0.2.0",
    "langchain-chroma==0.1.4",
    "langchain-cohere==0.3.0",
    "langchain-openai==0.2.0",
    "langgraph==0.2.22",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
    "ruff>=0.6.5",
    "mypy>=1.11.2",
]

[tool.ruff]
exclude = [".git", ".mypy_cache", ".ruff_cache", ".venv", "third_party"]
line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
line-ending = "auto"

[tool.ruff.lint]
fixable = ["ALL"]
unfixable = []
select = [
    "A",  # flake8-builtin
    "B",  # flake8-bugbear
    "E",  # pycodestyle error
    "F",  # Pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "W",  # pycodestyle warning
    "PL",  # Pylint
    "UP",  # pyupgrade
]
ignore = [
    "F401",  # Unused import
    "PLR0913",  # Too many arguments
    "PLR2004",  # Magic value used in comparison, consider replacing with a constant
    "PLW2901",  # `for` loop variable overwritten by assignment target
    "N803",  # Argument name should be lowercase
    "N806",  # Variable name should be lowercase
    "N812",  # Lowercase imported as non-lowercase
]
